+++ a.cpp
@@ -4,12 +4,12 @@ namespace RBPO {
 	namespace Lab2 {
 		namespace Variant15 {
 			namespace Task3 {
-				double a(unsigned short);
+				double a(long long);
 			};
 		};
 	};
 };
 
-double RBPO::Lab2::Variant15::Task3::a(unsigned short i) {
+double RBPO::Lab2::Variant15::Task3::a(long long i) {
 	return (i % 2 == 0 ? 1 : -1) * (i + 1) / double(i * i + 2 * i + 1);
 };
